% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a.r
\docType{methods}
\name{multi_method}
\alias{multi_method}
\alias{multi_method,ANY,ANY-method}
\alias{multi_method,numeric,ANY-method}
\alias{multi_method,numeric,integer-method}
\title{An s4 generic}
\usage{
multi_method(x, y)

\S4method{multi_method}{numeric,ANY}(x, y)

\S4method{multi_method}{numeric,integer}(x, y)
}
\description{
An s4 generic
}
